+++ packages/desktop/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @fusion/desktop
 
+## 0.73.0-beta.3
+
+### Patch Changes
+
+- @fusion/core@0.73.0-beta.3
+- @fusion/dashboard@0.73.0-beta.3
+- @fusion/engine@0.73.0-beta.3
+
+## 0.73.0-beta.2
+
+### Patch Changes
+
+- @fusion/core@0.73.0-beta.2
+- @fusion/dashboard@0.73.0-beta.2
+- @fusion/engine@0.73.0-beta.2
+
+## 0.73.0-beta.1
+
+### Patch Changes
+
+- @fusion/core@0.73.0-beta.1
+- @fusion/dashboard@0.73.0-beta.1
+- @fusion/engine@0.73.0-beta.1
+
+## 0.73.0-beta.0
+
+### Patch Changes
+
+- @fusion/core@0.73.0-beta.0
+- @fusion/dashboard@0.73.0-beta.0
+- @fusion/engine@0.73.0-beta.0
+
 ## 0.72.0
 
 ### Patch Changes